Invoke mentos directly on GNU/Linux
Invoking the script on GNU/Linux by passing it as an argument to
python(2) makes the process use 100% CPU for about tens of seconds when
it is supposed to timeout (test_returns_nil_on_timeout reproduces this).

On current Debian Sid python2 exists, so this can be used in mentos.py
shebang.

Brought back script invocation for windows.